[INTG-1618] Add extra column to sort table on
It looks like gorm is adding a default sort by the primary key to all queries. It thinks that ModifiedAt is the primary key as it is the only one present in the model.

Adding ExportedAt first makes this the default sort. It was chosen as it is present on all tables.