[INTG-1618] Fix exporting breaking when performing incremental exports against SQL Server (#47)
* [INTG-1618] Add failing integration test

* [INTG-1618] Add extra column to sort table on

It looks like gorm is adding a default sort by the primary key to all queries. It thinks that ModifiedAt is the primary key as it is the only one present in the model.

Adding ExportedAt first makes this the default sort. It was chosen as it is present on all tables.

* Add `workflow_dispatch` event trigger

* Remove shellcheck precommit

Something is going on with this precommit hook. I think it is caused by shellcheck being missing from the GitHub agent??

* Revert "Remove shellcheck precommit"

This reverts commit 6486d80b2e4356d98f944aff6aeb992a1bf5dbf8.

* Update ubuntu version

* Revert "Update ubuntu version"

This reverts commit 06f6557b7fb34eaddca85613a1abdfd455feb24b.

* Remove shellcheckers