[INTG-1617] Add flag for inspection modified after (#48)
* [INTG-1617] Add flag for inspection modified after

Adds a flag for `inspection.modified_after`. Also updates internal usages of ModifiedAfter to a Time instead of a string. This is done so that the time parsing of Viper is used instead of implementing it ourselves, viper supports a number of additonal formats now we don't need to implement!

* [INTG-1617] Use correct modified after

* [INTG-1617] Add supported formats to README